renames to node_locator_set.hpp
